+++ PSO-Shopkeeper/PSO-Shopkeeper-Lib/Item/EquippableItem.cs
@@ -106,7 +106,7 @@ namespace PSOShopkeeperLib.Item
         /// <summary>
         /// Indicates how much ATA this item provies, if applicable
         /// </summary>
-        public int ATA { get; set; } = 0;
+        public double ATA { get; set; } = 0.0;
 
         /// <summary>
         /// Indicates how much EVP this item provies, if applicable
